# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FIH Rules API",
    "description": "REST API for querying the official FIH (International Hockey Federation) rules for Outdoor, Indoor, and Hockey5s using RAG.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health Check",
        "description": "Check if the API and the RAG engine are initialized and healthy.\n\nReturns a 200 status if healthy, or 503 if the engine is not initialized.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Service health status",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "AI Chat",
        "description": "Process a chat request using the FIH Rules RAG engine.\n\nThis endpoint takes a user query and optional conversation history, reformulates the query \ninto a standalone question, performs a vector search over the FIH rules, and generates \nan answer using Gemini.",
        "operationId": "chat_chat_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI generated answer with source citations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evals/latest": {
      "get": {
        "tags": [
          "Evals"
        ],
        "summary": "Get Latest Metrics",
        "description": "Retrieve the most recent evaluation metrics generated by the synthetic evaluation pipeline.\n\nThe metrics include both custom calculations (accuracy, hit rate) and RAGAS scores \n(faithfulness, relevancy, etc.).",
        "operationId": "get_latest_evals_evals_latest_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest evaluation report metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "The user's question or query",
            "example": "What is a flick?"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "History",
            "description": "The conversation history for context",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "The AI-generated answer to the query"
          },
          "standalone_query": {
            "type": "string",
            "title": "Standalone Query",
            "description": "The query reformulated to be standalone based on history"
          },
          "variant": {
            "type": "string",
            "title": "Variant",
            "description": "The variant of the rules used (e.g., 'outdoor')"
          },
          "source_docs": {
            "items": {
              "$ref": "#/components/schemas/SourceDoc"
            },
            "type": "array",
            "title": "Source Docs",
            "description": "The source documents used to generate the answer"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "standalone_query",
          "variant",
          "source_docs"
        ],
        "title": "ChatResponse"
      },
      "EvalsMetrics": {
        "properties": {
          "custom_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Custom Metrics",
            "description": "Custom metrics calculated during evaluation (e.g., accuracy, hit rate)"
          },
          "ragas_metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ragas Metrics",
            "description": "RAGAS-specific metrics (e.g., faithfulness, answer relevancy)"
          }
        },
        "type": "object",
        "required": [
          "custom_metrics",
          "ragas_metrics"
        ],
        "title": "EvalsMetrics"
      },
      "EvalsResponse": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "ISO 8601 timestamp of when the evaluation report was generated"
          },
          "metrics": {
            "$ref": "#/components/schemas/EvalsMetrics",
            "description": "The evaluation metrics"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "metrics"
        ],
        "title": "EvalsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "description": "The role of the message sender (e.g., 'user', 'assistant')",
            "example": "user"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "The content of the message",
            "example": "What are the rules for a penalty corner?"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Message"
      },
      "SourceDoc": {
        "properties": {
          "page_content": {
            "type": "string",
            "title": "Page Content",
            "description": "The text content from the source document"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Metadata about the source document (e.g., source file, page number)"
          }
        },
        "type": "object",
        "required": [
          "page_content",
          "metadata"
        ],
        "title": "SourceDoc"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}